requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
drf-yasg==1.16.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.2
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.2
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.3
